+++ Changes.md
@@ -26,6 +26,7 @@ Development in progress.
 20. Fix 'Use of "$setting->set_updatedcallback('theme_reset_all_caches');" not needed on some settings' - #25.
 21. Fix 'lib.php preg_match logic flaw' - #150.
 22. Fix 'wrong rtl css' - #142.
+23. Fix 'Impossible to enter a course with Coventry tiles' - #156.
 
 Change Log in version 2.2.2 (2019112601)
 ========================================